Split AppConfigurationSettings.cs
